moved sd apis to a seperate file
